+++ bot/app.go
@@ -1,51 +1,20 @@
 package main
 
 import (
+	"ci6ndex-bot/domain"
 	"github.com/charmbracelet/log"
-	"github.com/spf13/viper"
 	"os"
 	"time"
 )
 
-type AppConfig struct {
-	DiscordToken     string `mapstructure:"DISCORD_API_TOKEN"`
-	BotApplicationID string `mapstructure:"DISCORD_BOT_APPLICATION_ID"`
-	GuildId          string `mapstructure:"CS_GUILD_ID"`
-	DatabaseUrl      string `mapstructure:"SQLITE_DB_URL"`
-}
-
-func loadConfig() (*AppConfig, error) {
-	var config AppConfig
-	viper.SetConfigFile(".env")
-	viper.SetConfigType("env")
-
-	viper.AutomaticEnv()
-
-	err := viper.ReadInConfig()
-	if err != nil {
-		return nil, err
-	}
-
-	err = viper.Unmarshal(&config)
-
-	if err != nil {
-		return nil, err
-	}
-	return &config, nil
-}
-
 type App struct {
-	db     *DatabaseOperations
+	db     *domain.DatabaseOperations
 	config *AppConfig
 	logger *log.Logger
 }
 
-func InitNewApp() (*App, error) {
-	config, err := loadConfig()
-	if err != nil {
-		panic(err)
-	}
-	db, err := newDBConnection(config.DatabaseUrl)
+func InitNewApp(c *AppConfig) (*App, error) {
+	db, err := domain.NewDBConnection(c.DatabaseUrl)
 	if err != nil {
 		panic(err)
 	}
@@ -57,7 +26,11 @@ func InitNewApp() (*App, error) {
 
 	return &App{
 		db:     db,
-		config: config,
+		config: c,
 		logger: logger,
 	}, nil
+}
+
+func (a *App) Serve() {
+	a.db.Close()
 }
