+++ bot/discord/discord.go
@@ -1,0 +1,11 @@
+package discord
+
+import (
+	"ci6ndex-bot/domain"
+	"github.com/bwmarrin/discordgo"
+)
+
+type Bot struct {
+	s  *discordgo.Session
+	db *domain.DatabaseOperations
+}
